requirements.txt
python-telegram-bot>=21.0
groq>=0.4.0
ddgs>=9.0.0
httpx>=0.27.0
feedparser>=6.0.10
pyyaml>=6.0.1
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
